requirements.txt
# # fastapi==0.110.0
# # uvicorn==0.19.0
# # pydantic==2.6.3
# # langchain==0.1.20
# # langchain-community==0.0.38
# # chromadb==0.4.24
# # sentence-transformers==2.2.2
# # huggingface-hub<1.0.0
# # torch
# # python-multipart
# # streamlit==1.37.0
# # requests>=2.31.0
# # python-dotenv==1.0.1
# # python-docx==0.8.11
# # pypdf==4.2.0
# # docx2txt



# # fastapi==0.110.0
# # uvicorn==0.22.0
# # python-multipart==0.0.6
# # pydantic==1.10.13
# # langchain==0.1.20
# # langchain-community==0.0.38
# # chromadb==0.4.24
# # sentence-transformers==2.2.2
# # huggingface-hub==0.13.4
# # torch==2.1.1
# # streamlit==1.37.0
# # python-dotenv==1.0.1
# # requests>=2.31.0
# # python-docx==0.8.11
# # pypdf==4.2.0
# # docx2txt

# # langchain==0.1.20
# # langchain-community==0.0.38
# # langchain-core==0.1.52
# # langsmith==0.1.17
# # langchain-text-splitters==0.0.2
# # langchain-chroma==0.0.4
# # huggingface-hub>=0.20.2
# # sentence-transformers>=2.7.0
# # chromadb>=0.4.24

# fastapi==0.110.0
# uvicorn==0.22.0
# python-multipart==0.0.6
# pydantic==1.10.13

# langchain==0.1.20
# langchain-community==0.0.38
# langchain-core==0.1.52
# langsmith==0.1.17
# langchain-text-splitters==0.0.2
# langchain-chroma==0.0.4

# chromadb==0.4.24
# sentence-transformers==2.7.0
# huggingface-hub==0.20.2
# torch==2.1.1

# streamlit==1.37.0
# python-dotenv==1.0.1
# requests>=2.31.0
# python-docx==0.8.11
# pypdf==4.2.0
# docx2txt


fastapi==0.110.0
#uvicorn>=0.22.0
pydantic==2.6.3
langchain==0.1.20
langchain-community==0.0.38
chromadb==0.4.24
sentence-transformers==2.2.2
huggingface-hub<1.0.0
torch
python-multipart
streamlit==1.37.0
requests>=2.31.0
python-dotenv==1.0.1
python-docx==0.8.11
pypdf==4.2.0
docx2txt
